Move addNewForm to its own file
